+++ common/src/main/kotlin/com/cobblemon/mod/common/api/events/pokemon/CollectEggEvent.kt
@@ -10,13 +10,12 @@ package com.cobblemon.mod.common.api.events.pokemon
 
 import com.cobblemon.mod.common.api.events.Cancelable
 import com.cobblemon.mod.common.api.pokemon.PokemonProperties
+import com.cobblemon.mod.common.pokemon.Pokemon
 import net.minecraft.server.level.ServerPlayer
 
-interface HatchEggEvent {
-    val egg: PokemonProperties
-    val player: ServerPlayer
-
-    data class Pre(override var egg : PokemonProperties, override var player: ServerPlayer) : HatchEggEvent, Cancelable()
-
-    data class Post(override var egg : PokemonProperties, override var player: ServerPlayer) : HatchEggEvent
-}
+class CollectEggEvent (
+    val egg : PokemonProperties,
+    val maleParent : Pokemon,
+    val femaleParent : Pokemon,
+    val player : ServerPlayer
+) : Cancelable()